# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda+ API",
        "description": "API RESTful para sistema de gestão de consultas médicas Agenda+",
        "contact": {
            "name": "Suporte Agenda+",
            "email": "suporte@agendaplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/appointments": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Listar consultas",
                "description": "Lista consultas do usuário autenticado. Retorna consultas diferentes conforme o perfil: paciente vê suas consultas, médico vê suas consultas, admin vê todas.",
                "operationId": "8aa3c0802355a81b59ed07d68f2f7867",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "CONFIRMED",
                                "COMPLETED",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Filtrar por período: future (futuras), past (passadas), all (todas)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "future",
                                "past",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de consultas paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Agendar consulta",
                "description": "Cria um novo agendamento de consulta. Requer antecedência mínima de 24h, perfil completo do paciente e médico ativo.",
                "operationId": "339467c94dffe0e8055dcc340f715673",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_id",
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-01 14:00:00"
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "PRESENTIAL"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Consulta de rotina"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta agendada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos ou regras de negócio violadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{id}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Obter detalhes da consulta",
                "description": "Retorna os detalhes completos de uma consulta incluindo médico, paciente, observações e logs",
                "operationId": "4e5d96ed4b005cf155a6af5d6eaac0cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da consulta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado para esta consulta"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{id}/confirm": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Confirmar consulta",
                "description": "Confirma uma consulta pendente. Apenas o médico responsável ou admin pode confirmar.",
                "operationId": "deed19c8fa711b1c1d4ad05aac275e1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta confirmada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Transição de status inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{id}/complete": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Marcar consulta como realizada",
                "description": "Marca uma consulta confirmada como realizada. Apenas médico ou admin pode executar.",
                "operationId": "0113ef4f8789ac15b5dbfde5108b76e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta marcada como realizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Transição de status inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{id}/cancel": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Cancelar consulta",
                "description": "Cancela uma consulta. Requer antecedência mínima de 12h. Pacientes e médicos podem cancelar suas próprias consultas, admin pode cancelar qualquer consulta.",
                "operationId": "7348b329f412a5316ba5b1789d91524a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Imprevisto pessoal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Cancelamento não permitido (menos de 12h de antecedência)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{id}/reschedule": {
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Remarcar consulta",
                "description": "Remarca uma consulta para nova data/horário. Máximo de 2 remarcações por consulta. Requer antecedência mínima de 12h.",
                "operationId": "00b8d2c694ff0a1c1c2e30e6f0120d1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-05 15:00:00"
                                    },
                                    "duration_minutes": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta remarcada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Limite de remarcações atingido ou dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}